requirements.txt
fastapi
uvicorn
python-dotenv
zadarma-api-v1
